+++ src/constants.ts
@@ -41,14 +41,14 @@ export const HomeFolder = IsWindows ? path.join(process.env['USERPROFILE'] || '.
 export const SystemBinFolder = IsWindows ? (process.env['SystemRoot'] || String.raw`C:\WINDOWS\system32`) : (IsMacOS ? '/usr/local/bin/' : '/usr/bin/');
 export const TempStorageFolder = IsWindows ? os.tmpdir() : '/tmp/';
 
-const GitInfoTemplate = "Skip_Git_Paths length = $L. Parsed $P patterns, omitted $E errors, ignored $X exemptions: see MSR-Def-Ref in OUTPUT tab.";
+const GitInfoTemplate = "Skip_Git_Paths length = $L. Parsed $P of $T patterns, omitted $E errors, ignored $X exemptions: see MSR-Def-Ref in OUTPUT tab.";
 
 export function getGitInfoTipTemplate(isCmdTerminal: boolean): string {
   return isCmdTerminal ? GitInfoTemplate.replace(/\$([A-Z])\b/g, '%$1%') : GitInfoTemplate; //.replace(/%([A-Z])%/, '$1')
 }
 
-export function getCommandToSetGitInfoVar(isCmdTerminal: boolean, skipGitRegexLength: number, parsedPatterns: number, errors: number, exemptions: number): string {
+export function getCommandToSetGitInfoVar(isCmdTerminal: boolean, skipGitRegexLength: number, totalPatterns: number, parsedPatterns: number, errors: number, exemptions: number): string {
   return isCmdTerminal
-    ? `set L=${skipGitRegexLength} & set P=${parsedPatterns} & set E=${errors} & set X=${exemptions} &`.replace(/ &/g, '&')
-    : `export L=${skipGitRegexLength}; export P=${parsedPatterns}; export E=${errors}; export X=${exemptions};`; //.replace(/export ([A-Z])/g, '$1');
+    ? `set L=${skipGitRegexLength} & set T=${totalPatterns} & set P=${parsedPatterns} & set E=${errors} & set X=${exemptions} &`.replace(/ &/g, '&')
+    : `export L=${skipGitRegexLength}; export T=${totalPatterns}; export P=${parsedPatterns}; export E=${errors}; export X=${exemptions};`; //.replace(/export ([A-Z])/g, '$1');
 }