+++ src/test/suite/extension.test.ts
@@ -6,7 +6,6 @@ import { checkConfigKeysInDoc, checkDuplicateDescription, validateRegexPatterns 
 import { testForLoopCmdAlias, testLinuxGeneralCmdAlias, testLinuxGeneralCmdAliasScript, testWindowsGeneralCmdAlias, testWindowsGeneralCmdAliasScript } from './cookCmdAliasTest';
 import { testCmdTerminalWithBackSlash, testCmdTerminalWithForwardSlash, testLinuxTerminal, testNotSkipDotPaths, testOmitExemptions } from './gitIgnoreTest';
 import { testEscapeRegex, testEscapeRegexForFindingCommands, testSpecialCaseReplacing } from './utilsTest';
-import { IsWindows } from '../../constants';
 
 suite('Test-1: Basic utils test', () => {
     before(() => {
@@ -56,10 +55,8 @@ suite('Test-3: Parsing .gitignore test', () => {
 });
 
 suite('Test-4: Cook each doskey/alias to a batch script file for Windows doskey', () => {
-    if (IsWindows) {
-        test('Test Windows doskey/alias of one file.', testWindowsGeneralCmdAlias);
-        test('Test Windows doskey/alias of multiple scripts.', testWindowsGeneralCmdAliasScript);
-    }
+    test('Test Windows doskey/alias of one file.', testWindowsGeneralCmdAlias);
+    test('Test Windows doskey/alias of multiple scripts.', testWindowsGeneralCmdAliasScript);
     test('Test Linux alias of multiple scripts.', testLinuxGeneralCmdAlias);
     test('Test Linux alias of multiple scripts.', testLinuxGeneralCmdAliasScript);
     test('Test use "%%x" for looping variable "%x" when cooking doskey/alias to files on Windows.', testForLoopCmdAlias);