+++ src/utils.ts
@@ -319,3 +319,8 @@ export function isPowerShellCommand(cmd: string, terminalType: TerminalType): bo
     const powerShellCmd = getPowerShellName(terminalType) + ' -Command';
     return cmd.includes(powerShellCmd);
 }
+
+export function getLoadAliasFileCommand(file: string, isWindowsTerminal: boolean, autoQuote: boolean = true): string {
+    const head = isWindowsTerminal ? 'doskey /MACROFILE=' : 'source ';
+    return head + (autoQuote ? quotePaths(file) : file);
+}